requirements.txt
pytest==5.4.1
tensorflow-gpu==1.15.0
pytorch
jupyter
tqdm==4.45.0